edit migrate field budgets_of_month to decimal

+++ database/migrations/2020_06_30_095928_create_budgets_table.php
@@ -15,7 +15,7 @@ class CreateBudgetsTable extends Migration
     {
         Schema::create('budgets', function (Blueprint $table) {
             $table->id();
-            $table->double('budgets_of_month');
+            $table->decimal('budgets_of_month',10,2);
             $table->string('month');
             $table->string('year');
             $table->timestamps();
